Migrate the Preconditions utility into a new package
This is a prereq for supporting java modules, previously preconditions
and the safe-logging API used the same package in separate jars, which
is not legal.

We must follow this up with an automated migration to the new
preconditions location, and updated error-prone checks to use
the new package name.

It's unfortunate that this will result in so much code churn,
but the migration should be relatively straightforward.