Migrate the Preconditions utility to its own package
Previously preconditions reused the same package as safe-logging
API, which is not allowed by the java platform module system.
While this change will impact a lot of code, we can automate the
transition in a relatively straightforward way, and eventualy shift
the legacy jar to a runtime dependency to prevent new uses.