Implement the `logger-log4j` module using log4j apis directly
This is meant to avoid slf4j indirection, and use a code-path
with better test coverage within log4j itself. For example,
this implementation will avoid the unintended allocations
incurred by LOG4J2-3560.

This module uses a bridge with dynamic priority based on the
presence of log4j-core. In the case that log4j-core is present,
it's preferred over the default.